refact: notifier load config when necessary
