+++ pkg/gateways/notifiers/stdout/stdout.go
@@ -1,10 +1,7 @@
 package stdout
 
 import (
-	"context"
-
 	"github.com/sirupsen/logrus"
-	"github.com/stone-co/webhook-consumer/pkg/common/configuration"
 	"github.com/stone-co/webhook-consumer/pkg/domain"
 )
 
@@ -16,17 +13,4 @@ type StdoutNotifier struct {
 
 func New() *StdoutNotifier {
 	return &StdoutNotifier{}
-}
-
-func (n *StdoutNotifier) Configure(config *configuration.Config, log *logrus.Logger) error {
-	n.log = log
-	return nil
-}
-
-func (n StdoutNotifier) Send(ctx context.Context, eventTypeHeader, eventIDHeader, body string) error {
-	log := n.log.WithField("notifier", "stdout")
-	log.Printf("event headers: type[%s] id[%s]\n", eventTypeHeader, eventIDHeader)
-	log.Printf("body: %s\n", body)
-
-	return nil
 }
